Safari-611.1.10/LayoutTests/editing/deleting/backspace-at-table-cell-beginning.html

<!DOCTYPE html>
<html>
<body>
<p id="description">This test verifies that backspace at the beginning of a table-cell is a no-op. The first dump verifies that the caret stays before 'abc', and doesn't select the nested table. The second dump verifies that the caret stays before 'def'.</p>
<div id="test" contenteditable="true"><table><tr><td><table><tr><td>ghi</td><td>jkl</td></tr></table></td><td id="first">abc</td><td id="second">def</td></tr></table></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
Markup.description(document.getElementById("description").innerText);

window.getSelection().collapse(document.getElementById("first").firstChild, 0);
document.execCommand("Delete");
Markup.dump("test");

window.getSelection().collapse(document.getElementById("second").firstChild, 0);
document.execCommand("Delete");
Markup.dump("test");
</script>
</body>
</html>
